+++ src/main/java/com/study/loge/betchat/controller/HomeController.java
@@ -1,6 +1,6 @@
 package com.study.loge.betchat.controller;
 
-import com.study.loge.betchat.model.requests.UserLoginRequest;
+import com.study.loge.betchat.model.request.UserLoginRequest;
 import com.study.loge.betchat.model.response.UserLoginResponse;
 import com.study.loge.betchat.model.MessageHeader;
 import com.study.loge.betchat.service.UserLoginService;
