+++ src/main/java/com/study/loge/betchat/exception/RoomCreateException.java
@@ -1,4 +1,4 @@
-package com.study.loge.betchat.exceptions;
+package com.study.loge.betchat.exception;
 
 // 채팅방 생성 시 발생하는 예외입니다.
 public class RoomCreateException extends Exception {
