+++ src/main/java/com/study/loge/betchat/exception/UserLoginException.java
@@ -1,4 +1,4 @@
-package com.study.loge.betchat.exceptions;
+package com.study.loge.betchat.exception;
 
 // 유저 로그인 단계에서 발생하는 예외입니다.
 public class UserLoginException extends Exception {
