+++ src/main/java/com/study/loge/betchat/service/RoomCreateService.java
@@ -4,9 +4,9 @@ import com.study.loge.betchat.component.KeyGenerator;
 import com.study.loge.betchat.entity.Room;
 import com.study.loge.betchat.entity.User;
 import com.study.loge.betchat.enums.ResultState;
-import com.study.loge.betchat.exceptions.RoomCreateException;
+import com.study.loge.betchat.exception.RoomCreateException;
 import com.study.loge.betchat.model.MessageHeader;
-import com.study.loge.betchat.model.requests.RoomCreateRequest;
+import com.study.loge.betchat.model.request.RoomCreateRequest;
 import com.study.loge.betchat.model.response.RoomCreateResponse;
 import com.study.loge.betchat.repository.RoomRepository;
 import com.study.loge.betchat.repository.UserRepository;
